// src/maps/tarrain-enum.ts
/* eslint-disable no-unused-vars */
export enum Terrain {
  // Row 1
  r1c1 = '001',
  r1c2 = '002',
  r1c3 = '003',
  r1c4 = '004',
  r1c5 = '005',
  r1c6 = '006',
  r1c7 = '007',
  r1c8 = '008',
  r1c9 = '009',
  r1c10 = '010',
  // Row 2
  r2c1 = '011',
  r2c2 = '012',
  r2c3 = '013',
  r2c4 = '014',
  r2c5 = '015',
  r2c6 = '016',
  r2c7 = '017',
  r2c8 = '018',
  r2c9 = '019',
  r2c10 = '020',
  // Row 3
  r3c1 = '021',
  r3c2 = '022',
  r3c3 = '023',
  r3c4 = '024',
  r3c5 = '025',
  r3c6 = '026',
  r3c7 = '027',
  r3c8 = '028',
  r3c9 = '029',
  r3c10 = '030',
  // Row 4
  r4c1 = '031',
  r4c2 = '032',
  r4c3 = '033',
  r4c4 = '034',
  r4c5 = '035',
  r4c6 = '036',
  r4c7 = '037',
  r4c8 = '038',
  r4c9 = '039',
  r4c10 = '040',
  // Row 5
  r5c1 = '041',
  r5c2 = '042',
  r5c3 = '043',
  r5c4 = '044',
  r5c5 = '045',
  r5c6 = '046',
  r5c7 = '047',
  r5c8 = '048',
  r5c9 = '049',
  r5c10 = '050',
  // Row 6
  r6c1 = '051',
  r6c2 = '052',
  r6c3 = '053',
  r6c4 = '054',
  r6c5 = '055',
  r6c6 = '056',
  r6c7 = '057',
  r6c8 = '058',
  r6c9 = '059',
  r6c10 = '060',
  // Row 7
  r7c1 = '061',
  r7c2 = '062',
  r7c3 = '063',
  r7c4 = '064',
  r7c5 = '065',
  r7c6 = '066',
  r7c7 = '067',
  r7c8 = '068',
  r7c9 = '069',
  r7c10 = '070',
  // Row 8
  r8c1 = '071',
  r8c2 = '072',
  r8c3 = '073',
  r8c4 = '074',
  r8c5 = '075',
  r8c6 = '076',
  r8c7 = '077',
  r8c8 = '078',
  r8c9 = '079',
  r8c10 = '080',
  // Row 9
  r9c1 = '081',
  r9c2 = '082',
  r9c3 = '083',
  r9c4 = '084',
  r9c5 = '085',
  r9c6 = '086',
  r9c7 = '087',
  r9c8 = '088',
  r9c9 = '089',
  r9c10 = '090',
  // Row 10
  r10c1 = '091',
  r10c2 = '092',
  r10c3 = '093',
  r10c4 = '094',
  r10c5 = '095',
  r10c6 = '096',
  r10c7 = '097',
  r10c8 = '098',
  r10c9 = '099',
  r10c10 = '100',
  // Row 11
  r11c1 = '101',
  r11c2 = '102',
  r11c3 = '103',
  r11c4 = '104',
  r11c5 = '105',
  r11c6 = '106',
  r11c7 = '107',
  r11c8 = '108',
  r11c9 = '109',
  r11c10 = '110',
  // Row 12
  r12c1 = '111',
  r12c2 = '112',
  r12c3 = '113',
  r12c4 = '114',
  r12c5 = '115',
  r12c6 = '116',
  r12c7 = '117',
  r12c8 = '118',
  r12c9 = '119',
  r12c10 = '120',
  // Row 13
  r13c1 = '121',
  r13c2 = '122',
  r13c3 = '123',
  r13c4 = '124',
  r13c5 = '125',
  r13c6 = '126',
  r13c7 = '127',
  r13c8 = '128',
  r13c9 = '129',
  r13c10 = '130',
  // Row 14
  r14c1 = '131',
  r14c2 = '132',
  r14c3 = '133',
  r14c4 = '134',
  r14c5 = '135',
  r14c6 = '136',
  r14c7 = '137',
  r14c8 = '138',
  r14c9 = '139',
  r14c10 = '140',
  // Row 15
  r15c1 = '141',
  r15c2 = '142',
  r15c3 = '143',
  r15c4 = '144',
  r15c5 = '145',
  r15c6 = '146',
  r15c7 = '147',
  r15c8 = '148',
  r15c9 = '149',
  r15c10 = '150',
  // Row 16
  r16c1 = '151',
  r16c2 = '152',
  r16c3 = '153',
  r16c4 = '154',
  r16c5 = '155',
  r16c6 = '156',
  r16c7 = '157',
  r16c8 = '158',
  r16c9 = '159',
  r16c10 = '160',
  // Row 17
  r17c1 = '161',
  r17c2 = '162',
  r17c3 = '163',
  r17c4 = '164',
  r17c5 = '165',
  r17c6 = '166',
  r17c7 = '167',
  r17c8 = '168',
  r17c9 = '169',
  r17c10 = '170',
  // Row 18
  r18c1 = '171',
  r18c2 = '172',
  r18c3 = '173',
  r18c4 = '174',
  r18c5 = '175',
  r18c6 = '176',
  r18c7 = '177',
  r18c8 = '178',
  r18c9 = '179',
  r18c10 = '180',
  // Row 19
  r19c1 = '181',
  r19c2 = '182',
  r19c3 = '183',
  r19c4 = '184',
  r19c5 = '185',
  r19c6 = '186',
  r19c7 = '187',
  r19c8 = '188',
  r19c9 = '189',
  r19c10 = '190',
  // Row 20
  r20c1 = '191',
  r20c2 = '192',
  r20c3 = '193',
  r20c4 = '194',
  r20c5 = '195',
  r20c6 = '196',
  r20c7 = '197',
  r20c8 = '198',
  r20c9 = '199',
  r20c10 = '200',
  // Row 21
  r21c1 = '201',
  r21c2 = '202',
  r21c3 = '203',
  r21c4 = '204',
  r21c5 = '205',
  r21c6 = '206',
  r21c7 = '207',
  r21c8 = '208',
  r21c9 = '209',
  r21c10 = '210',
  // Row 22
  r22c1 = '211',
  r22c2 = '212',
  r22c3 = '213',
  r22c4 = '214',
  r22c5 = '215',
  r22c6 = '216',
  r22c7 = '217',
  r22c8 = '218',
  r22c9 = '219',
  r22c10 = '220',
}